pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "looker-migrator"
version = "0.1.0"
description = "Looker LookML to Power BI TMDL migration tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.14"
authors = [
    {name = "Codehive", email = "dev@codehive.com"}
]
keywords = [
    "looker",
    "lookml",
    "power-bi",
    "tmdl",
    "migration",
    "bi-tools",
    "data-modeling",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pybars3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "hatchling",
]
compile = [
    "requests>=2.28",
    "websocket-client>=1.5",
]
all = [
    "looker-migrator[dev,compile]",
]

[project.scripts]
looker-migrator = "looker_migrator.main:main"

[project.urls]
Homepage = "https://github.com/Codehive-Inc/Bi-Migrator-Looker"
Documentation = "https://github.com/Codehive-Inc/Bi-Migrator-Looker#readme"
Repository = "https://github.com/Codehive-Inc/Bi-Migrator-Looker.git"
Issues = "https://github.com/Codehive-Inc/Bi-Migrator-Looker/issues"

[tool.hatch.build.targets.wheel]
packages = ["looker_migrator"]

[tool.hatch.build.targets.sdist]
include = [
    "looker_migrator/*",
    "*.yaml",
    "*.json",
    "*.md",
    "*.txt",
    "templates/*",
    "config/*"
]

[dependency-groups]
dev = [
    "hatchling",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | \.eggs
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["looker_migrator"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
